// src/app/sections/claim-section/claim/conversation/to-timeline-info/model/timeline-action.ts
export enum TimelineAction {
    ChangesAdded = 'changesAdded',
    FilesAdded = 'filesAdded',
    FilesDeleted = 'filesDeleted',
    CommentAdded = 'commentAdded',
    StatusReview = 'statusReview',
    StatusPending = 'statusPending',
    StatusDenied = 'statusDenied',
    StatusRevoked = 'statusRevoked',
    StatusAccepted = 'statusAccepted',
}